==> 02_describe.Rmd <==
---
title: "01_test"
author: "Cal Chengqi Fang"
date: "2/19/2022"
output: pdf_document
---

```{r setup, include=FALSE}
# Clear out the history
rm(list=ls())

# Read in the packages needed
library(tidyverse)
library(rmarkdown)
library(knitr)
library(ggplot2)
library(scales)
library(gmodels)
library(weights)

# Set working directory
setwd("~/Box Sync/CCSSBP Project")    # Please change this to your path

# STEP 1
# Read in the dataset
load("data/gp.Rdata")
# Take out the variables needed for describing
# Also filter out records of graduate students
gp_describe <- gp %>% 
  select(c(1, 3:32, 65:66, 210:233, scityl, weight)) %>% 
  filter(sttype == 1 | sttype == 2) %>% 
  filter(syear %in% c(2019, 2018, 2017, 2016)) %>% 
  rename(slevel = name)
```

# 1. Sample Description
## 1.1 Sampling Strategy
The dataset we will use in this project is the outcome of the NCSS-SRH survey project. This project used a combination of multi-stage random sampling method and snowball convenience sampling method to collect the data. 

In the first stage, researchers built the sampling frame based on the distribution of higher education schools in the eastern, central, and western regions, referring to the directory of higher education institutions published by the Ministry of Education of China in June 2018. Due to the severely uneven distribution of higher education institutions in China at the provincial (administrative region level equal to states in America) level (e.g., there are only seven colleges and universities in the whole Tibetan region), there are difficulties in sampling at the provincial level, and the stratification of the sample colleges and universities in the first sampling stage was performed using the East, Central and West divisions based on administrative divisions provided by the National Bureau of Statistics. 

In the second stage, colleges offering associate's or bachelor's degree were selected from each region. Selected colleges offering bachelor's degrees were then divided into four levels in each region: first-class universities, universities of first-class disciplines, general undergraduate colleges and universities, and private undergraduate colleges and universities. Institutions offering associate's degrees were divided into three levels: key associate's colleges and universities, general associate's colleges and universities, and private associate's colleges and universities. Corresponding number of universities are selected according to the proportion of each level in the overall, with an appropriate inclination to the colleges and universities with more geographically diverse student source, taking into account the capability of the survey implementation organization, China Youth Network. A total of 241 sample colleges and universities are selected. The purpose of sampling at school level in this way is to ensure that the social networks in the study have nationwide first-level dissemination points, and to ensure that college students in 31 provinces and regions nationwide have access to the questionnaire through social media and can be included in the sample (in fact, college students in 34 provincial-level administrative regions, including Hong Kong, Macao and Taiwan, are included in the snowballing process). However, the samples from Hong Kong, Macao and Taiwan were suggested to be excluded in the analysis stage because of the significant differences in social background, cultural environment and education system between these samples and those from mainland China. 

The 241 sample universities were used as the starting point for snowballing to sample college students enrolled in higher education institutions nationwide. All samples were *volunteer samples* recruited in the form of convenience sampling, with students deciding whether or not to participate in the survey and being able to withdraw at any time during the process after participation, with full disclosure of survey ethics given in the guidelines.    

Referring to the [literature review](https://docs.google.com/spreadsheets/d/1v9XOB-lS4rFisT7RmA-kim8wW2pZCGIyJu7bzNWwmGo/edit?usp=sharing), we could see that most of the similar existing researches on youth sexual knowledge, attitudes, and behavior conducted in other Asian countries used random samples. However, except some researches done in Thailand benefiting from the nationwide military mandate lottery, the random samples used in existing researches mostly were generated from a relatively small pool, for instance, 12 schools (Hedayati-Moghaddam et al., 2015), 1 community (Jaya & Hindin, 2009), or 3 factories (Tang et al., 2011). In that sense, although the random selection of our data might be limited due to adopting snowball sampling, the comparatively large size might be able to compensate and ensure some representativeness.  
    
## 1.2 Sample Structure
Because of the limited size of graduate students sample, I excluded all graduate student samples here. (P.S. 0 in the variable gender refers to female students while 1 refers to male students.)
    
### 1.2.1 Enrollment Year       
```{r, warning=FALSE, message=FALSE, echo=FALSE}
table(gp_describe$gender, gp_describe$syear) %>% 
  kable()
```
As above shows, for both females and males, this dataset has more sample of students enrolled in 2019, who are the first-year college students at the time of investigation. There doesn't seem to exist much variation between the distributions of enrollment year of males and females.    
    
### 1.2.2 City Level
```{r, warning=FALSE, message=FALSE, echo=FALSE}
prop.table(table(gp_describe$gender, gp_describe$scityl), 1) %>% 
  kable()
```
The female sample distributed comparatively even among level 1 to 5 cities, while the male samples seem more concentrated in the level 2 and level 3 cities.       

### 1.2.3 School Tier   
```{r, warning=FALSE, message=FALSE, echo=FALSE}
prop.table(table(gp_describe$gender, gp_describe$slevel), 1) %>% 
  kable()
```
There are more female samples in every school tier level except 1st level associate's degree institutions. 

### 1.2.4 Sex Orientation
```{r, warning=FALSE, message=FALSE, echo=FALSE}
prop.table(table(gp_describe$gender, gp_describe$seorient), 1) %>% 
  kable()
prop.table(table(gp_describe$slevel, gp_describe$seorient), 1) %>% 
  kable()
```
Above is the distribution of sex orientation over gender and school tiers.            
   
   
## 2. Draft Tests on Penerative Sex
### 2.1 Cross-tables and chisq-tests
#### 2.1.1 City Level   


```{r, warning=FALSE, message=FALSE, echo=FALSE}
gp_f <- gp_describe %>% 
  filter(gender == 0,
         !is.na(as.numeric(sbpenage)),
         as.numeric(sbpenage) > 12,
         as.numeric(sbpenage) < 26)
gp_m <- gp_describe %>% 
  filter(gender == 1,
         !is.na(as.numeric(sbpenage)),
         as.numeric(sbpenage) > 12,
         as.numeric(sbpenage) < 26)
prop.table(table(gp_f$scityl, round(as.numeric(gp_f$sbpenage))), 1) %>% 
  round(2) %>% 
  kable()
```
The proportion table above shows the distribution of first penetrative sex age of female students in different level cities. ***The row proportion distribution actually seems pretty similar across different levels.***
```{r, warning=FALSE, message=FALSE, echo=FALSE}
prop.table(table(gp_f$scityl, round(as.numeric(gp_f$sbpenage))), 1) %>% 
  round(2) %>% 
  kable()
```
The table above shows the distribution of first penetrative sex age of male students in different level cities. ***The row proportion distribution also seems pretty similar across different levels below 20, while above 20 there seems to be some difference across different city level.***               

#### 2.1.2 School Tier        
```{r, warning=FALSE, message=FALSE, echo=FALSE}
prop.table(table(gp_f$slevel, round(as.numeric(gp_f$sbpenage))), 1) %>% 
  round(2) %>% 
  kable()
```
Above is the distribution of first penetrative sex age of female students in different level schools. ***It seems like female students in associate's degree institutions tend to have sex at a younger age (below 19).***        
```{r, warning=FALSE, message=FALSE, echo=FALSE}
prop.table(table(gp_m$slevel, round(as.numeric(gp_m$sbpenage))), 1) %>% 
  round(2) %>% 
  kable()
```
Above is the distribution of first penetrative sex age of male students in different level schools. ***Same as above, male students in associate's degree institutions also tend to have sex at a younger age (below 19).***         

# 3 Distribution of the sex orientation variables
The three distribution tables are as follows:       

```{r, warning=FALSE, message=FALSE, echo=FALSE}
gp_describe$gender <- as.factor(gp_describe$gender)
levels(gp_describe$gender) <- c("Female", "Male")
prop.table(table(gp_describe$gender)) %>% 
  kable(col.names = c("Biological Gender", "Frequency"),
        caption = "the Distribution of Biological Gender")
```

```{r, warning=FALSE, message=FALSE, echo=FALSE}
gp_describe$seorient <- as.factor(gp_describe$seorient)
levels(gp_describe$seorient) <- c("Heterosexual", "Homosexual", "Bisexual", "Pansexual", "Asexual", "Others", "Unsure")
prop.table(table(gp_describe$seorient)) %>% 
  kable(col.names = c("Sexual Orientation", "Frequency"),
        caption = "the Distribution of Sexual Orientation")
```

```{r, warning=FALSE, message=FALSE, echo=FALSE}
prop.table(table(gp_describe$gender, gp_describe$seorient)) %>% 
  kable(caption = "the Cross-distribution of Biological Gender")
```
